Moved temperature to shared
