+++ Content.Client/Temperature/Systems/EntityHeaterSystem.cs
@@ -1,0 +1,7 @@
+﻿using Content.Shared.Temperature.Systems;
+
+namespace Content.Client.Temperature.Systems;
+
+public sealed class EntityHeaterSystem : SharedEntityHeaterSystem
+{
+}
