+++ Content.Server/Body/Systems/ThermalRegulatorSystem.cs
@@ -1,17 +1,11 @@
-using Content.Server.Body.Components;
-using Content.Server.Temperature.Components;
-using Content.Server.Temperature.Systems;
-using Content.Shared.ActionBlocker;
-using Robust.Shared.Timing;
+using Content.Shared.Body.Components;
+using Content.Shared.Body.Systems;
+using Content.Shared.Temperature.Components;
 
 namespace Content.Server.Body.Systems;
 
-public sealed class ThermalRegulatorSystem : EntitySystem
+public sealed class ThermalRegulatorSystem : SharedThermalRegulatorSystem
 {
-    [Dependency] private readonly IGameTiming _gameTiming = default!;
-    [Dependency] private readonly TemperatureSystem _tempSys = default!;
-    [Dependency] private readonly ActionBlockerSystem _actionBlockerSys = default!;
-
     public override void Initialize()
     {
         base.Initialize();
