+++ Content.Server/Temperature/Systems/EntityHeaterSystem.cs
@@ -1,32 +1,16 @@
-using Content.Server.Power.Components;
-using Content.Server.Temperature.Components;
-using Content.Shared.Examine;
+﻿using Content.Server.Power.Components;
 using Content.Shared.Placeable;
-using Content.Shared.Popups;
 using Content.Shared.Temperature;
-using Content.Shared.Verbs;
-using Robust.Server.Audio;
+using Content.Shared.Temperature.Components;
+using Content.Shared.Temperature.Systems;
 
 namespace Content.Server.Temperature.Systems;
 
-/// <summary>
-/// Handles <see cref="EntityHeaterComponent"/> updating and events.
-/// </summary>
-public sealed class EntityHeaterSystem : EntitySystem
+public sealed class EntityHeaterSystem : SharedEntityHeaterSystem
 {
-    [Dependency] private readonly SharedAppearanceSystem _appearance = default!;
-    [Dependency] private readonly SharedPopupSystem _popup = default!;
-    [Dependency] private readonly TemperatureSystem _temperature = default!;
-    [Dependency] private readonly AudioSystem _audio = default!;
-
-    private readonly int SettingCount = Enum.GetValues(typeof(EntityHeaterSetting)).Length;
-
     public override void Initialize()
     {
         base.Initialize();
-
-        SubscribeLocalEvent<EntityHeaterComponent, ExaminedEvent>(OnExamined);
-        SubscribeLocalEvent<EntityHeaterComponent, GetVerbsEvent<AlternativeVerb>>(OnGetVerbs);
         SubscribeLocalEvent<EntityHeaterComponent, PowerChangedEvent>(OnPowerChanged);
     }
 
@@ -44,38 +28,9 @@ public sealed class EntityHeaterSystem : EntitySystem
             var energy = power.PowerReceived * deltaTime;
             foreach (var ent in placer.PlacedEntities)
             {
-                _temperature.ChangeHeat(ent, energy);
+                Temperature.ChangeHeat(ent, energy);
             }
         }
-    }
-
-    private void OnExamined(EntityUid uid, EntityHeaterComponent comp, ExaminedEvent args)
-    {
-        if (!args.IsInDetailsRange)
-            return;
-
-        args.PushMarkup(Loc.GetString("entity-heater-examined", ("setting", comp.Setting)));
-    }
-
-    private void OnGetVerbs(EntityUid uid, EntityHeaterComponent comp, GetVerbsEvent<AlternativeVerb> args)
-    {
-        if (!args.CanAccess || !args.CanInteract)
-            return;
-
-        var setting = (int) comp.Setting;
-        setting++;
-        setting %= SettingCount;
-        var nextSetting = (EntityHeaterSetting) setting;
-
-        args.Verbs.Add(new AlternativeVerb()
-        {
-            Text = Loc.GetString("entity-heater-switch-setting", ("setting", nextSetting)),
-            Act = () =>
-            {
-                ChangeSetting(uid, nextSetting, comp);
-                _popup.PopupEntity(Loc.GetString("entity-heater-switched-setting", ("setting", nextSetting)), uid, args.User);
-            }
-        });
     }
 
     private void OnPowerChanged(EntityUid uid, EntityHeaterComponent comp, ref PowerChangedEvent args)
@@ -83,32 +38,22 @@ public sealed class EntityHeaterSystem : EntitySystem
         // disable heating element glowing layer if theres no power
         // doesn't actually turn it off since that would be annoying
         var setting = args.Powered ? comp.Setting : EntityHeaterSetting.Off;
-        _appearance.SetData(uid, EntityHeaterVisuals.Setting, setting);
+        Appearance.SetData(uid, EntityHeaterVisuals.Setting, setting);
     }
 
-    private void ChangeSetting(EntityUid uid, EntityHeaterSetting setting, EntityHeaterComponent? comp = null, ApcPowerReceiverComponent? power = null)
+    public override void ChangeSetting(Entity<EntityHeaterComponent?> heater, EntityHeaterSetting setting)
     {
-        if (!Resolve(uid, ref comp, ref power))
+        ChangeSetting((heater, heater.Comp, null), setting);
+    }
+
+    public void ChangeSetting(Entity<EntityHeaterComponent?, ApcPowerReceiverComponent?> heater,
+        EntityHeaterSetting setting)
+    {
+        base.ChangeSetting(heater, setting);
+        if (!Resolve(heater,ref heater.Comp2))
             return;
-
-        comp.Setting = setting;
-        power.Load = SettingPower(setting, comp.Power);
-        _appearance.SetData(uid, EntityHeaterVisuals.Setting, setting);
-        _audio.PlayPvs(comp.SettingSound, uid);
-    }
-
-    private float SettingPower(EntityHeaterSetting setting, float max)
-    {
-        switch (setting)
-        {
-            case EntityHeaterSetting.Low:
-                return max / 3f;
-            case EntityHeaterSetting.Medium:
-                return max * 2f / 3f;
-            case EntityHeaterSetting.High:
-                return max;
-            default:
-                return 0f;
-        }
+        heater.Comp2.Load = SettingPower(setting, heater.Comp1!.Power);
+        Appearance.SetData(heater, EntityHeaterVisuals.Setting, setting);
+        Audio.PlayPvs(heater.Comp1.SettingSound, heater);
     }
 }
