+++ Content.Shared/Body/Components/ThermalRegulatorComponent.cs
@@ -1,10 +1,11 @@
-using Content.Server.Body.Systems;
+using Content.Shared.Body.Systems;
+using Robust.Shared.GameStates;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom;
 
-namespace Content.Server.Body.Components;
+namespace Content.Shared.Body.Components;
 
-[RegisterComponent]
-[Access(typeof(ThermalRegulatorSystem))]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
+[Access(typeof(SharedThermalRegulatorSystem))]
 public sealed partial class ThermalRegulatorComponent : Component
 {
     /// <summary>
@@ -22,43 +23,43 @@ public sealed partial class ThermalRegulatorComponent : Component
     /// <summary>
     /// Heat generated due to metabolism. It's generated via metabolism
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float MetabolismHeat;
 
     /// <summary>
     /// Heat output via radiation.
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float RadiatedHeat;
 
     /// <summary>
     /// Maximum heat regulated via sweat
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float SweatHeatRegulation;
 
     /// <summary>
     /// Maximum heat regulated via shivering
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float ShiveringHeatRegulation;
 
     /// <summary>
     /// Amount of heat regulation that represents thermal regulation processes not
     /// explicitly coded.
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float ImplicitHeatRegulation;
 
     /// <summary>
     /// Normal body temperature
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float NormalBodyTemperature;
 
     /// <summary>
     /// Deviation from normal temperature for body to start thermal regulation
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float ThermalRegulationTemperatureThreshold;
 }
