+++ Content.Shared/Temperature/Components/TemperatureComponent.cs
@@ -2,58 +2,59 @@ using Content.Shared.Alert;
 using Content.Shared.Atmos;
 using Content.Shared.Damage;
 using Content.Shared.FixedPoint;
+using Robust.Shared.GameStates;
 using Robust.Shared.Prototypes;
 
-namespace Content.Server.Temperature.Components;
+namespace Content.Shared.Temperature.Components;
 
 /// <summary>
 /// Handles changing temperature,
 /// informing others of the current temperature,
 /// and taking fire damage from high temperature.
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
 public sealed partial class TemperatureComponent : Component
 {
     /// <summary>
     /// Surface temperature which is modified by the environment.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float CurrentTemperature = Atmospherics.T20C;
 
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float HeatDamageThreshold = 360f;
 
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float ColdDamageThreshold = 260f;
 
     /// <summary>
     /// Overrides HeatDamageThreshold if the entity's within a parent with the TemperatureDamageThresholdsComponent component.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float? ParentHeatDamageThreshold;
 
     /// <summary>
     /// Overrides ColdDamageThreshold if the entity's within a parent with the TemperatureDamageThresholdsComponent component.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float? ParentColdDamageThreshold;
 
     /// <summary>
     /// Heat capacity per kg of mass.
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float SpecificHeat = 50f;
 
     /// <summary>
     /// How well does the air surrounding you merge into your body temperature?
     /// </summary>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public float AtmosTemperatureTransferEfficiency = 0.1f;
 
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public DamageSpecifier ColdDamage = new();
 
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public DamageSpecifier HeatDamage = new();
 
     /// <summary>
@@ -62,18 +63,18 @@ public sealed partial class TemperatureComponent : Component
     /// <remarks>
     /// Okay it genuinely reaches this basically immediately for a plasma fire.
     /// </remarks>
-    [DataField, ViewVariables(VVAccess.ReadWrite)]
+    [DataField, ViewVariables(VVAccess.ReadWrite), AutoNetworkedField]
     public FixedPoint2 DamageCap = FixedPoint2.New(8);
 
     /// <summary>
     /// Used to keep track of when damage starts/stops. Useful for logs.
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public bool TakingDamage;
 
-    [DataField]
+    [DataField, AutoNetworkedField]
     public ProtoId<AlertPrototype> HotAlert = "Hot";
 
-    [DataField]
+    [DataField, AutoNetworkedField]
     public ProtoId<AlertPrototype> ColdAlert = "Cold";
 }
